kafka/consumer/requirements.txt
confluent-kafka==2.3.0
boto3==1.34.0
pyarrow==15.0.0
python-dotenv==1.0.0